+++ client/src/pages/Stats/StatsDashboard.js
@@ -1,9 +1,11 @@
 import React, { Component } from "react";
 import { axiosWithAuth, baseURL } from "../../config/axiosWithAuth";
+import moment from "moment";
 
 import PageTitle from "../../components/PageTitle";
 import SummaryBox from "../../components/SummaryBox";
 
+import { InputLabel, MenuItem, FormControl, Select } from "@material-ui/core";
 import CircularProgress from "@material-ui/core/CircularProgress";
 
 import "./StatsDashboard.css";
@@ -16,7 +18,10 @@ class StatsDashboard extends Component {
     super();
     this.state = {
       reports: [],
-      users: []
+      users: [],
+      data: [],
+      labels: [],
+      filterBy: ""
     };
   }
 
@@ -70,6 +75,49 @@ class StatsDashboard extends Component {
       .catch(err => console.log(err));
   };
 
+  generateDates = num => {
+    let generatedDates = [];
+    for (let i = num; i > 0; i--) {
+      let date = new Date();
+      date.setDate(date.getDate() - i);
+      date = moment(date).format("l");
+      generatedDates.push(date);
+    }
+    this.setState({
+      labels: generatedDates
+    });
+  };
+
+  handleChange = e => {
+    this.setState({
+      [e.target.name]: e.target.value
+    });
+  };
+
+  setLabels = e => {
+    e.preventDefault();
+    switch (this.state.filterBy) {
+      case "day":
+        let today = moment(Date.now()).format("l");
+        this.setState({
+          labels: [today]
+        });
+        break;
+      case "week":
+        this.generateDates(7);
+        break;
+      case "month":
+        this.generateDates(30);
+        break;
+      case "quarter":
+        this.generateDates(90);
+        break;
+      case "year":
+        this.generateDates(365);
+        break;
+    }
+  };
+
   render() {
     if (this.state.reports.length === 0) {
       return (
@@ -97,7 +145,42 @@ class StatsDashboard extends Component {
               content={this.state.reports.length}
             />
           </div>
-          <ChartOptions reports={this.state.reports} />
+          <form
+            autoComplete="off"
+            className="form"
+            style={{ display: "flex", flexWrap: "wrap" }}
+          >
+            <FormControl className="formControl">
+              <InputLabel
+                className="inputLabel"
+                style={{ marginBottom: "20px" }}
+              >
+                Filter By
+              </InputLabel>
+
+              <Select
+                value={this.state.filterBy}
+                style={{ marginTop: "20px" }}
+                name="filterBy"
+                onChange={this.handleChange}
+                className="select"
+              >
+                <MenuItem value={"day"}>Day</MenuItem>
+                <MenuItem value={"week"}>Week</MenuItem>
+                <MenuItem value={"month"}>Month</MenuItem>
+                <MenuItem value={"quarter"}>Quarter</MenuItem>
+                <MenuItem value={"year"}>Year</MenuItem>
+              </Select>
+              <button onClick={this.setLabels} style={{ margin: "20px" }}>
+                Filter
+              </button>
+            </FormControl>
+          </form>
+          <ChartOptions
+            reports={this.state.reports}
+            data={this.state.data}
+            labels={this.state.labels}
+          />
         </div>
         <div className="sidebar" />
       </div>
